fix: add ticket to login api  & implement logout api
